Content update
Added a strong tag and modified it to be consistent with the website's
look. Increased padding to each entry. Hid scrolling for the overall
body, so that users don't see two scroll bars when the width of the
window becomes less than the width of the blog entry. Also adding a
min-width to the header bar so the circle icon on the top left does not
overlap with it.